chore: move files for directory strategy
