+++ src/pages/FamilyEntryPage.tsx
@@ -1,10 +1,10 @@
 import { useEffect } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import lucky1 from '../../assets/lucky/level1.svg';
-import Button from '../../components/common/Button';
-import { ServiceConstant } from '../../constants/service';
-import useModalStore from '../../store/modalStore';
-import FamilyCodeModal from './components/FamilyCodeModal';
+import lucky1 from '../assets/lucky/level1.svg';
+import Button from '../components/common/Button';
+import FamilyCodeModal from '../components/familyentry/FamilyCodeModal';
+import { ServiceConstant } from '../constants/service';
+import useModalStore from '../store/modalStore';
 
 const FamilyEntryPage = () => {
   const navigate = useNavigate();
@@ -20,18 +20,18 @@ const FamilyEntryPage = () => {
 
   useEffect(() => {
     const code = searchParams.get('code');
-    
+
     if (code)
-    openModal({
-      content: <FamilyCodeModal code={code}/>,
-      showCloseBtn: false,
-    })
+      openModal({
+        content: <FamilyCodeModal code={code} />,
+        showCloseBtn: false,
+      });
   }, []);
 
   return (
     <div className="flex h-screen w-full flex-col items-center justify-between px-12 py-10 text-center">
       <div>
-        <p className="font-ryurue text-ryurue-lg mt-[120px]">{SERVICE_NAME}</p>
+        <p className="mt-[120px] font-ryurue text-ryurue-lg">{SERVICE_NAME}</p>
         <p className="font-ryurue text-ryurue-base">{WELCOME_MESSAGE}</p>
       </div>
       <img src={lucky1} alt="lucky-ground" className="w-4/5" />
