+++ src/routes.tsx
@@ -5,9 +5,9 @@ import routes from './constants/routes';
 import ChallengDetailPage from './pages/ChallengDetailPage';
 import ChallengeListPage from './pages/ChallengeListPage';
 import FamilyCreatePage from './pages/FamilyCreatePage';
-import FamilyEntryPage from './pages/FamilyEntryPage/FamilyEntryPage';
+import FamilyEntryPage from './pages/FamilyEntryPage';
 import LoginPage from './pages/LoginPage';
-import MainPage from './pages/MainPage/MainPage';
+import MainPage from './pages/MainPage';
 import MyPage from './pages/MyPage';
 import RecapPage from './pages/RecapPage';
 import SignupPage from './pages/SignupPage';
